+++ src/main/java/org/synyx/urlaubsverwaltung/application/application/ApplicationForLeaveDetailsViewController.java
@@ -115,7 +115,7 @@ class ApplicationForLeaveDetailsViewController implements HasLaunchpad {
     }
 
     @GetMapping("/{applicationId}")
-    public String showApplicationDetail(@PathVariable("applicationId") Integer applicationId,
+    public String showApplicationDetail(@PathVariable("applicationId") Long applicationId,
                                         @RequestParam(value = "year", required = false) Integer requestedYear,
                                         @RequestParam(value = "action", required = false) String action,
                                         @RequestParam(value = "shortcut", required = false) boolean shortcut, Model model)
@@ -143,7 +143,7 @@ class ApplicationForLeaveDetailsViewController implements HasLaunchpad {
      */
     @PreAuthorize(IS_BOSS_OR_DEPARTMENT_HEAD_OR_SECOND_STAGE_AUTHORITY)
     @PostMapping("/{applicationId}/allow")
-    public String allowApplication(@PathVariable("applicationId") Integer applicationId,
+    public String allowApplication(@PathVariable("applicationId") Long applicationId,
                                    @ModelAttribute("comment") ApplicationCommentForm comment, Errors errors,
                                    @RequestParam(value = "redirect", required = false) String redirectUrl,
                                    RedirectAttributes redirectAttributes) throws UnknownApplicationForLeaveException {
@@ -198,7 +198,7 @@ class ApplicationForLeaveDetailsViewController implements HasLaunchpad {
      */
     @PreAuthorize(IS_PRIVILEGED_USER)
     @PostMapping("/{applicationId}/refer")
-    public String referApplication(@PathVariable("applicationId") Integer applicationId,
+    public String referApplication(@PathVariable("applicationId") Long applicationId,
                                    @ModelAttribute("referredPerson") ReferredPerson referredPerson, RedirectAttributes redirectAttributes)
         throws UnknownApplicationForLeaveException, UnknownPersonException {
 
@@ -228,7 +228,7 @@ class ApplicationForLeaveDetailsViewController implements HasLaunchpad {
 
     @PreAuthorize(IS_BOSS_OR_DEPARTMENT_HEAD_OR_SECOND_STAGE_AUTHORITY)
     @PostMapping("/{applicationId}/reject")
-    public String rejectApplication(@PathVariable("applicationId") Integer applicationId,
+    public String rejectApplication(@PathVariable("applicationId") Long applicationId,
                                     @ModelAttribute("comment") ApplicationCommentForm comment, Errors errors,
                                     @RequestParam(value = "redirect", required = false) String redirectUrl,
                                     RedirectAttributes redirectAttributes) throws UnknownApplicationForLeaveException {
@@ -276,7 +276,7 @@ class ApplicationForLeaveDetailsViewController implements HasLaunchpad {
      * Cancelling an application for leave on behalf for someone is allowed only for Office.
      */
     @PostMapping("/{applicationId}/cancel")
-    public String cancelApplication(@PathVariable("applicationId") Integer applicationId,
+    public String cancelApplication(@PathVariable("applicationId") Long applicationId,
                                     @ModelAttribute("comment") ApplicationCommentForm comment, Errors errors,
                                     RedirectAttributes redirectAttributes)
         throws UnknownApplicationForLeaveException {
@@ -323,7 +323,7 @@ class ApplicationForLeaveDetailsViewController implements HasLaunchpad {
      * Cancel the cancellation request of an application for leave.
      */
     @PostMapping("/{applicationId}/decline-cancellation-request")
-    public String declineCancellationRequestApplication(@PathVariable("applicationId") Integer applicationId,
+    public String declineCancellationRequestApplication(@PathVariable("applicationId") Long applicationId,
                                                         @ModelAttribute("comment") ApplicationCommentForm comment, Errors errors,
                                                         RedirectAttributes redirectAttributes)
         throws UnknownApplicationForLeaveException {
@@ -358,7 +358,7 @@ class ApplicationForLeaveDetailsViewController implements HasLaunchpad {
      * Remind the bosses about the decision of an application for leave.
      */
     @PostMapping("/{applicationId}/remind")
-    public String remindBoss(@PathVariable("applicationId") Integer applicationId,
+    public String remindBoss(@PathVariable("applicationId") Long applicationId,
                              RedirectAttributes redirectAttributes) throws UnknownApplicationForLeaveException {
 
         final Application application = applicationService.getApplicationById(applicationId)
