+++ src/main/java/org/synyx/urlaubsverwaltung/application/application/ApplicationServiceImpl.java
@@ -42,7 +42,7 @@ class ApplicationServiceImpl implements ApplicationService {
     }
 
     @Override
-    public Optional<Application> getApplicationById(Integer id) {
+    public Optional<Application> getApplicationById(Long id) {
         return applicationRepository.findById(id);
     }
