+++ src/main/java/org/synyx/urlaubsverwaltung/application/specialleave/SpecialLeaveSettingsItem.java
@@ -2,19 +2,19 @@ package org.synyx.urlaubsverwaltung.application.specialleave;
 
 public class SpecialLeaveSettingsItem {
 
-    private final Integer id;
+    private final Long id;
     private final Boolean active;
     private final String messageKey;
     private final Integer days;
 
-    public SpecialLeaveSettingsItem(Integer id, Boolean active, String messageKey, Integer days) {
+    public SpecialLeaveSettingsItem(Long id, Boolean active, String messageKey, Integer days) {
         this.id = id;
         this.active = active;
         this.messageKey = messageKey;
         this.days = days;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
