+++ src/main/java/org/synyx/urlaubsverwaltung/application/specialleave/SpecialLeaveSettingsRepository.java
@@ -4,11 +4,11 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-interface SpecialLeaveSettingsRepository extends CrudRepository<SpecialLeaveSettingsEntity, Integer> {
+interface SpecialLeaveSettingsRepository extends CrudRepository<SpecialLeaveSettingsEntity, Long> {
 
     @Override
     List<SpecialLeaveSettingsEntity> findAll();
 
     @Override
-    List<SpecialLeaveSettingsEntity> findAllById(Iterable<Integer> ids);
+    List<SpecialLeaveSettingsEntity> findAllById(Iterable<Long> ids);
 }
