+++ src/main/java/org/synyx/urlaubsverwaltung/application/vacationtype/VacationTypeRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-interface VacationTypeRepository extends JpaRepository<VacationTypeEntity, Integer> {
+interface VacationTypeRepository extends JpaRepository<VacationTypeEntity, Long> {
 
     List<VacationTypeEntity> findByActiveIsTrue();
 }
