+++ src/main/java/org/synyx/urlaubsverwaltung/notification/UserNotificationSettingsEntity.java
@@ -10,14 +10,14 @@ import java.util.Objects;
 class UserNotificationSettingsEntity {
 
     @Id
-    private Integer personId;
+    private Long personId;
     private boolean restrictToDepartments;
 
-    public Integer getPersonId() {
+    public Long getPersonId() {
         return personId;
     }
 
-    public void setPersonId(Integer personId) {
+    public void setPersonId(Long personId) {
         this.personId = personId;
     }
